Authentication and Authorization

+++ controllers/api/v1/posts_api.js
@@ -27,8 +27,8 @@ module.exports.destroy = async (req, res) => {
   try {
     let post = await Post.findById(req.params.id);
     // .id means converting the object id into string
-    // if (post.user == req.user.id) {
-      // console.log(post);
+    if (post.user == req.user.id) {
+      console.log(post);
       post.remove();
       await Comment.deleteMany({ post: req.params.id });
 
@@ -37,10 +37,11 @@ module.exports.destroy = async (req, res) => {
       return res.json(200,{
         message: "Post and associated comments deleted successfully"
       });
-    // } else {
-    //   req.flash('error','You can not delete this post!');
-    //   return res.redirect("back");
-    // }
+    } else {
+        return res.json(401, {
+          message: "You can not delete this post! " 
+        })
+    }
   } catch (err) {
     console.log('*********',err);
     return res.json(500,{
